add getListOrders

+++ auth/Model/Order.cs
@@ -6,7 +6,7 @@
         public string UserName { get; set; }
         public string Address { get; set; }
         public string Phone { get; set; }
-        public float Total { get; set; }
+        public int Total { get; set; }
         public bool IsApproved { get; set; } = false;
         public string PaymentMethod { get; set; }
         public DateTime PaymentTime { get; set; }
